+++ app/GraphQL/Controllers/User/LikeController.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace App\GraphQL\Controllers\User;
 
+use App\Exceptions\GraphQL\ClientValidationException;
 use App\GraphQL\Controllers\BaseController;
-use App\GraphQL\Definition\Mutations\User\LikeMutation;
-use App\GraphQL\Definition\Mutations\User\UnlikeMutation;
+use App\GraphQL\Definition\Mutations\Models\User\LikeMutation;
+use App\GraphQL\Definition\Mutations\Models\User\UnlikeMutation;
 use App\Models\List\Playlist;
 use App\Models\User\Like;
 use App\Models\Wiki\Video;
-use Exception;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Arr;
 
@@ -30,7 +30,7 @@ class LikeController extends BaseController
      * @param  null  $_
      * @param  array  $args
      *
-     * @throws Exception
+     * @throws ClientValidationException
      */
     public function store($_, array $args): Model
     {
@@ -51,7 +51,7 @@ class LikeController extends BaseController
             return $video;
         }
 
-        throw new Exception('None models detected to like.');
+        throw new ClientValidationException('One resource is required to like.');
     }
 
     /**
@@ -60,7 +60,7 @@ class LikeController extends BaseController
      * @param  null  $_
      * @param  array  $args
      *
-     * @throws Exception
+     * @throws ClientValidationException
      */
     public function destroy($_, array $args): Model
     {
@@ -81,6 +81,6 @@ class LikeController extends BaseController
             return $video;
         }
 
-        throw new Exception('None models detected to unlike.');
+        throw new ClientValidationException('One resource is required to unlike.');
     }
 }
