+++ app/GraphQL/Definition/Mutations/Models/List/Playlist/DeletePlaylistMutation.php
@@ -2,16 +2,18 @@
 
 declare(strict_types=1);
 
-namespace App\GraphQL\Definition\Mutations\Rest\List\Playlist;
+namespace App\GraphQL\Definition\Mutations\Models\List\Playlist;
 
 use App\GraphQL\Attributes\Resolvers\UseFieldDirective;
 use App\GraphQL\Controllers\List\PlaylistController;
-use App\GraphQL\Definition\Mutations\Rest\UpdateMutation;
+use App\GraphQL\Definition\Mutations\Models\DeleteMutation;
 use App\GraphQL\Definition\Types\List\PlaylistType;
+use App\GraphQL\Definition\Types\MessageResponseType;
 use App\Models\List\Playlist;
+use GraphQL\Type\Definition\Type;
 
-#[UseFieldDirective(PlaylistController::class, 'update')]
-class UpdatePlaylistMutation extends UpdateMutation
+#[UseFieldDirective(PlaylistController::class, 'destroy')]
+class DeletePlaylistMutation extends DeleteMutation
 {
     public function __construct()
     {
@@ -23,7 +25,7 @@ class UpdatePlaylistMutation extends UpdateMutation
      */
     public function description(): string
     {
-        return 'Update playlist';
+        return 'Delete playlist';
     }
 
     /**
@@ -32,5 +34,13 @@ class UpdatePlaylistMutation extends UpdateMutation
     public function baseType(): PlaylistType
     {
         return new PlaylistType();
+    }
+
+    /**
+     * The type returned by the field.
+     */
+    public function getType(): Type
+    {
+        return Type::nonNull(new MessageResponseType());
     }
 }
