+++ app/GraphQL/Definition/Mutations/Models/List/Playlist/Track/UpdatePlaylistTrackMutation.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace App\GraphQL\Definition\Mutations\Rest\List\Playlist\Track;
+namespace App\GraphQL\Definition\Mutations\Models\List\Playlist\Track;
 
 use App\GraphQL\Attributes\Resolvers\UseFieldDirective;
 use App\GraphQL\Controllers\List\Playlist\PlaylistTrackController;
-use App\GraphQL\Definition\Mutations\Rest\UpdateMutation;
+use App\GraphQL\Definition\Mutations\Models\UpdateMutation;
 use App\GraphQL\Definition\Types\List\Playlist\PlaylistTrackType;
 use App\Models\List\Playlist\PlaylistTrack;
