+++ app/GraphQL/Definition/Mutations/Models/UpdateMutation.php
@@ -2,23 +2,25 @@
 
 declare(strict_types=1);
 
-namespace App\GraphQL\Definition\Mutations\Rest;
+namespace App\GraphQL\Definition\Mutations\Models;
 
+use App\Contracts\GraphQL\Fields\UpdatableField;
 use App\Contracts\GraphQL\HasFields;
 use App\GraphQL\Definition\Argument\Argument;
+use App\GraphQL\Definition\Fields\Field;
 use App\GraphQL\Definition\Mutations\BaseMutation;
 use GraphQL\Type\Definition\Type;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Arr;
 
-abstract class DeleteMutation extends BaseMutation
+abstract class UpdateMutation extends BaseMutation
 {
     /**
      * @param  class-string<Model>  $model
      */
     public function __construct(protected string $model)
     {
-        parent::__construct('delete'.ucfirst(class_basename($model)));
+        parent::__construct('update'.ucfirst(class_basename($model)));
     }
 
     /**
@@ -34,6 +36,7 @@ abstract class DeleteMutation extends BaseMutation
 
         if ($baseType instanceof HasFields) {
             $arguments[] = $this->resolveBindArguments($baseType->fields());
+            $arguments[] = $this->resolveUpdateMutationArguments($baseType->fields());
         }
 
         return Arr::flatten($arguments);
@@ -48,13 +51,33 @@ abstract class DeleteMutation extends BaseMutation
     {
         return [
             'canModel' => [
-                'ability' => 'delete',
+                'ability' => 'update',
                 'injectArgs' => true,
                 'model' => $this->model,
             ],
 
             ...parent::directives(),
         ];
+    }
+
+    /**
+     * Get the rules for the create mutation.
+     *
+     * @param  array<string, mixed>  $args
+     * @return array<string, array>
+     */
+    public function rules(array $args): array
+    {
+        $baseType = $this->baseType();
+
+        if ($baseType instanceof HasFields) {
+            return collect($baseType->fields())
+                ->filter(fn (Field $field) => $field instanceof UpdatableField)
+                ->mapWithKeys(fn (Field&UpdatableField $field) => [$field->getColumn() => $field->getUpdateRules($args)])
+                ->toArray();
+        }
+
+        return [];
     }
 
     /**
